added a solvent's CoolProp name to solvent data
'name_in_coolprop' is added to the SolventData as a new attribute.
CoolProp is an external packaged used for fluid property calculations.
CoolProp takes a solvent's name as an input, which is saved as
'name_in_coolprop' for each solvent.
For solvents that are not available in CoolProp, 'name_in_coolprop' is
set to None