Modified main.py to allow ranged based for liquid phase reactors.
Checked if the reaction_system.T exists. If so, the viscosity will be calculated. If not, then it's range based reactor; the viscosity will be calculated in each sim using temperature stored in conditions (conditions['T']).
Checked conditions is not None before calculating viscosity with conditions['T'] to avoid breaking the non-ranged based liquid reactor.