Change dictionary iterator to avoid RuntimeError
The py3 version dict.keys() gives an iterator for the dictionary's keys, so it would have "RuntimeError: dictionary changed size during iteration" since we have del template_rxn_map[L] here. The py2 version's dict.keys() gives a copy of dictionary's key so it didn't have this problem. Here we can use list(dict.keys()) or list(dict) to force it to return a copy of the dictionary's keys.