Log the problem reaction if it causes an error making a species.
If there's a problem making a species or estimating its thermochemistry
it can be helpful to know where the species came from to help debugging.
This will report the reaction that was being processed when the species was made
and where it came from, e.g. the kinetics library, seed mechanism, or reaction family.
Then it will raise the original error as if nothing happened.
Helps address #2030.