Fix format specifier when logging species pruning.
Hopefully this closes #2043

The Species class doesn't have an ability to interpret a format
specifier "<56" (although the Species.label would, because it's a string.

Anyway, while fixing that I also changed some other 
logging.debug() statements over to the old-fashioned format syntax
and left the formatting to the logging module. This way, it formats 
the string "lazily", and if you're not running in verbose mode (i.e.
the debug messages aren't going to be printed anyway) then it doesn't
bother doing the formatting. Might save a few nanoseconds :-)

Because we almost always run with logging.info() showing, I left those alone.