Constrain conda channels.
Resolving the conda dependency tree takes a long time (>800 sec on Travis).
This is an attempt to speed up the solution by removing some channel options
and specifying which packages come from RMG channel.

The specifications are based on a recent travis run, and may be over-specified
for other OS systems.