Trick cclib parser into returning the rotcons and molmass
The parsing of the rotational constants and the molar mass that Greg Magoon
put into the RMG version of the MOPAC parser did make its way into the official
version, but these attributes are not included in the default set returned by
all parsers. If you make the attributes on the parser object, however, then they
are extracted and preserved, and you can then put them on to the data object.

A bit of a hack, but works. We could/should suggest/request that cclib makes these 
proper attributes that are dealt with the same way as all others.