Use secret for MOPAC license key.
This is stored as an organization level secret.
For security reasons, it won't be passed to actions triggered
by a pull request from a forked repository.
But then it'll just run the unit tests without MOPAC.

We also now cope with SIGPIPE error installing MOPAC.
The goal is to get the MOPAC step to exit with exit code 0 (no error).

When mopac finishes, the pipe gets closed, and the `yes` tool
then fails with a SIGPIPE error because it is piping 'Yes' to
a closed pipe. With "bash -o pipefail" (the default in Github Actions)
the piped commands will fail if any of them fails, so it fails because
of the failure code of yes. So we set "+o pipefail" to counteract
this, so that it has a nonzero exit code only if the mopac step fails.