If the cclib_data has no molmass, calculate it from the atoms.
If the cclib_data object returned by the cclib parsing of the QM
calculation doesn't include a molar mass (this is not included
by default in the official cclib version) then calculate it from
the atomic numbers. Notice that the values don't exactly agree with
those used by MOPAC, so you can be off by a fraction of a percent.
Also, this won't cope with unusual isotopes.