Added molecule.has_halogen, atom.is_halogen methods
Added a new function to atom class, is_halogen(), which returns True if
the atom is F, Cl, Br, or I and False if not.
A new function is also added to molecule class, has_halogen(), which
returns True if a molecule has at least one halogen atom (F, Cl, Br, I)
or returns False if not.

The corresponding unittests are created in moleculeTest.py.