Added halogen correction for solute data
Halogen correction is applied to solute data in a similar way
as the radical HBI correction.

For `get_solute_data` method, if the original molecule is not found in
the solute libraries, there are two cases to consider:

First case: the species is not a radical and contains halogen atoms.
If the species contains halogen atoms but is not a radical, the halogen
atoms are replaced with hydrogen atoms first and the method checks
whether the replaced form can be found in the solute libraries. If yes,
is uses the solute data of the replaced form and applies a halogen
correction for each halogenated site on the original molecule.
If the replaced form is not found in the solute libraries, the solute data
is computed for the replaced structure using the group additivity. Then
halogen correction is applied for each halogenated site on the original
molecule.

Second case: the species is a radical and contains halogen atoms.
If the species is a radical and contains halogen atoms, the radical
sites are saturated first and `get_solute_data` method checks whether the
saturated form can be found in the solute libraries. If yes, it uses the
solute data of the saturated from and applies the HBI correction on the
original molecule.
If the saturated form cannot be found in the solute libraries, it then
replaces the halogen atoms with hydrogen atoms and checks whether the
saturated, replaced form can be found in the solute libraries. If yes,
it uses the solute data of the saturated, replaces form, applies a
halogen correction for each halogenated site on the saturated from, and
finally performs a HBI correction on the original molecule.
If the saturated, replaced form cannot be found in the solute libraries,
it computes the solute data for the saturated, replaced form using the
group additivity, applies a halogen correction for each halogenated site
on the saturated form, and finally performs a HBI correction on the
original molecule.